[feat#261] 알림 api 구현

+++ Backend/Heroes/src/main/java/com/hq/heroes/notification/entity/NotificationCategory.java
@@ -15,7 +15,7 @@ public class NotificationCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "notification_category_id")
-    private int notificationCategoryId;
+    private Long notificationCategoryId;
 
     @Column(name = "category_name")
     private String notificationCategoryName;
